Support for legacy docker registry warning message
After the Docker 1.8 release and the new Docker Registry 2.1, the docker API returns an unexpected message for Ansible when pulling a docker image if the docker registry hasn't been upgraded. Although the pull is successful, Ansible fails because the warning message below was not considered in lines 1406 - 1407:

"<docker_image>: this image was pulled from a legacy registry.  Important: This registry version will not be supported in future versions of docker."

In order to make Ansible capable to accept the above message as a successful pull, I suggest to consider this message as a successful pull.

More info:

The old version of the docker registry: https://github.com/docker/docker-registry.
The neo version of the docker registry: https://github.com/docker/distribution